yay, we now have markers on map ,they should still get clickables

+++ server/db/schema/02_restaurants.sql
@@ -8,7 +8,7 @@ CREATE TABLE restaurants (
   start_time TIME,
   end_time TIME,
   address TEXT,
-  lat INTEGER,
-  lng INTEGER
+  lat DECIMAL(8,6),
+  lng DECIMAL(8,6)
 );
